requirements.txt
typing_extensions
requests
streamlit
pandas
numpy
fastapi
python-dotenv
mangum
fhir.resources